more fcn shuffling
