+++ pvlib/bifacial/utils.py
@@ -208,7 +208,7 @@ def vf_ground_sky_2d_integ(surface_tilt, surface_azimuth, gcr, height,
         Integration of view factor over the length between adjacent, interior
         rows.  Shape matches that of ``surface_tilt``. [unitless]
     """
-    # Abuse utils._vf_ground_sky_2d by supplying surface_tilt in place
+    # Abuse vf_ground_sky_2d by supplying surface_tilt in place
     # of a signed rotation. This is OK because
     # 1) z span the full distance between 2 rows, and
     # 2) max_rows is set to be large upstream, and
